+++ stats.jl
@@ -23,12 +23,12 @@ end
 
 
 # adjusted rsquared
-function adjustedr2score(yvec, ŷvec, p::Int)
+function adjustedr2score(yvec, ŷvec, p::Int16)
     M = size(yvec, 1)
     return 1 - (1 - r2score(yvec, ŷvec)) * (M - 1) / (M - p - 1)
 end
 
-function adjustedr2score_multidim(ys, ŷs, p::Int, multioutput::String="uniformaverage")
+function adjustedr2score_multidim(ys, ŷs, p::Int16, multioutput::String="uniformaverage")
     d = size(ys, 2)
     adjustedr2score_rawvalues = [adjustedr2score(ys[:, i], ŷs[:, i], p) for i in 1:d]
     if multioutput == "rawvalues"
@@ -62,7 +62,7 @@ function initscoresdict(n_folds; include_losses=false)
 end
 
 
-function updatescoresdict!(scores_dict, fold_id, y_train, y_train_preds, y_val, y_val_preds, n_features::Int, training_losses=nothing)
+function updatescoresdict!(scores_dict, fold_id, y_train, y_train_preds, y_val, y_val_preds, n_features::Int16, training_losses=nothing)
     scores_dict["r2score_train"][fold_id] = r2score(y_train, y_train_preds)
     scores_dict["r2score_val"][fold_id] = r2score(y_val, y_val_preds)
